update benchmark to use new c api

+++ examples/benchmark/build.rs
@@ -7,9 +7,13 @@ use tar::Archive;
 
 fn main() {
     // Clone the original C encoder/decoder repository
-    let c_path = "./qoi_c";
-    let _ = std::fs::remove_dir_all(c_path);
-    let _ = Repository::clone("https://github.com/phoboslab/qoi", c_path).unwrap();
+    {
+        if File::open("./qoi_c/qoi.h").is_err() {
+            let c_path = "./qoi_c";
+            let _ = std::fs::remove_dir_all(c_path);
+            let _ = Repository::clone("https://github.com/phoboslab/qoi", c_path).unwrap();
+        }
+    }
 
     // Compile and link the C files
     cc::Build::new()
@@ -20,7 +24,7 @@ fn main() {
 
     // Download the PNG suite tarball
     {
-        if !File::open("./img/images.tar").is_ok() {
+        if File::open("./img/images.tar").is_err() {
             let _ = std::fs::create_dir("./img");
             let mut downloader = Downloader::builder()
                 .download_folder(Path::new("./img"))
@@ -35,7 +39,7 @@ fn main() {
     }
 
     // Extract the PNG suite tarball
-    if !File::open("./img/images/kodak/kodim01.png").is_ok() {
+    if File::open("./img/images/kodak/kodim01.png").is_err() {
         let mut archive = Archive::new(File::open("./img/images.tar").unwrap());
         archive.unpack("./img").unwrap();
     }
